Fix ZipFileListener not initializing
The plugin lifecycle is:
- EventListener.__init__()
- plugin_loaded()
- EventListener.on_init(views)

Thus API functions like sublime.active_window().active_widget() can't be
used in ZipFileListener.__init__() to check if an overlay panel is
open. In fact, doing so silently prevents the event listener from
initializing.

Instead it must be assumed that no overlay panel is open when the event
listener is initialized. And on version 4050 and later,
ZipFileListener.on_init(views) is used to check if the overlay panel is
open.